Eliminated the on-off button. Multiple threads were all checking if the on-off button had been set to "off", but this caused a problem: the control thread would only set it to "off" after it had seen the number of failed tests had exceed the configured maximum. In the time between checking if the maximum had been exceeded and hitting the "off" button, thousands of other tests were already executed and had dumped their report files. Now, the 'onOff' function checks if the number of failed tests exceeds the maximum, and if so, returns false right away.
Created an exception handler 'tryTest' to be used to catch exceptions for individual sub-tests in
the 'checkTestCase' function. As a result, exceptions occurring in the check function are now reported
as a failed test in the report for that individual test case. Before the exception would not be
caught until the 'loopOnInput' function which would dump the exception without any in-depth information
about the test case right into the "uncaught.log" file.

The parser test now removes location information and comments from the parsed AST object before
comparing it to the original randomly-generated object. Before, when it had not done this, the
comparison function would evaluate to false for identical AST data structures merely because the
location information was included in the structure returned by the parser. To do this, I had to
introduce the 'delLocation' function. I also introduced the 'Dao.Object.AST.canonicalize' function
into the 'Dao.Object.AST.Intermediate' class.

The act of creating the 'delLocation' function introduced an avalanche of other changes which will
be committed soon.