The token and location data types are fundamental data types because they are used to construct the abstract syntax tree and the data types that define the Dao interpreted language. So it is necessary to keep these data types in a separate module so the "Dao.Evaluator", "Dao.Object" and "Dao.Parser" can all make use of them without causing circular dependencies.
Up until now, the old parser had its token and location types were defined here, and in order to coexist with the new parser, the new parser had it's own location and token types defined within the "Dao.NewParser" module.

However now that I am ready to being automated testing, I cannot have two different sets of location and token types. Well, actually I could have had two different sets, but it is no longer worth the trouble to keep the old parser up to date with all the changes made to the data types upon which it depends. The new parser is becoming stable enough, it is clearly better, and I do not intend to go back to the old parser, so I am phasing it out now.

To begin transitioning the code base to make of the permanent use of the new parser, I have deleted all the code in the "Dao.Token" module and moved most of the location and token-related data types from the "Dao.NewParser" module to the "Dao.Token" module.