I changed the type of string constant in the 'Dao.Glob.Glob' expressions used throughout the evaluator from 'Dao.String.UStr' to 'Dao.Evaluator.Object'. This means arbitrary Dao object values can be used to match to strings. Since 'Dao.Evaluator.Object's can overload the equality function, non-string values can be programmed to be used in rule expressions as well.
Therefore I have moved the 'FuzzyStr' type from "Dao.Glob" to "Dao.Evaluator": this type will be made a built-in type so fuzzy string matching can be implemented as soon as it becomes possible to define which function to use for string tokenizing. This has not been implemented yet.

I have restored the 'Dao.Evaluator.programTokenizer' to the 'Dao.Evaluator.ExecUnit' data structure, but unlike before it is now a 'Dao.Evaluator.Object'. The 'Dao.Evaluator.programTokenizer' will need to be set on execution of the 'Dao.Evaluator.Program' data type, but this has not been done yet, nor have the functions which tokenize rules and input strings been made to use anything other than 'Dao.String.simpleTokenizer'.

Also, I fixed a few bugs in the "Dao.Glob" module I had not caught yesterday.