Defined a new function 'Dao.Stack.stackUpdateTopM'. The problem with the function 'Dao.Stack.stackUpdateM' is that when you want to bind a variable that has not been bound before, it will scan every level of the stack looking for a bound variable name, and update it if it exists.  That means if you push a new context onto the stack and bind a variable name to a variable name bound in a higher level, the already-bound variable will be re-bound. When the stack is popped, that modification remains because it occurred in a higher level of the stack. This is not how programmers expect name binding to work when binding a previously unbound variable name using the equal "=" operator. It should only work this way when looking-up a value already bound to a variable name.
The 'Dao.Stack.stackUpdateTopM' function works as expected: When assigning values to unbound variable names will always occur in the top level of the stack, even if the name is bound at a lower level. This behavior is more similar to how C-like languages work, where variables names bound within any block of code enclosed in curly-brackets will mask identical variable names in it's containing block of code. The 'Dao.Stack.stackUpdateM' function can still be used when updating a value for an already-bound variable name and not binding an unbound variable name.