Corrected two bugs uncovered by the test routine.
On line 708, in the parser for the logical not prefix operator "!", the order of parsing whitespace and the expression following the "import" or "require" keyword had been reversed so any expression where whitespace occurred between the "!" operator and the operand was backtracking, resulting in a parser error.

On line 827, there is a parser that parses 'Dao.Evaluator.AST_Script' expressions that must inspect object expressions to determine whether or not they must be followed by a semicolon. Function and rule expressions may be followed by a semicolon. If they are followed by a semicolon they form 'Dao.Evaluator.AST_Eval' expressions rather than 'AST_RuleFunc' expressions -- this is how the pretty-printer differentiates between whether or not a semicolon was used. However the 'Dao.Evaluator.AST_Script' parser that inspected whether or not a semicolon was used formed a parser that was masking the 'Dao.Evaluator.AST_Arith' expressions parser. Therefore if an 'Dao.Evaluator.AST_Arith' expression used a function or rule expression on the left-hand side of an infix operator the expression would be parsed as a 'Dao.Evaluator.AST_RuleFunc' expression, and then fail on the infix operator as the 'Dao.Evaluator.AST_RuleFunc' parser returned control to the 'Dao.Evaluator.AST_Script' expression parser which expecting another 'Dao.Evaluator.AST_Script' expression rather than an infix operator. This has now been fixed.