I have tried and tried, but I just cannot keep the "Dao.Object" and "Dao.Evaluator" modules separate. There is really no one dependency that I can cut that will cleanly separate the two.
The problem is all of the data types form a very knotted circular dependency loop, which looks something like this:

Object --> HaskellData --> Interface
Interface --> Exec --> ExecUnit
Interface --> CallableCode --> Subroutine
ExecUnit --> Subroutine
Subroutine --> CodeBlock --> ScriptExpr
ScriptExpr --> (everything else) --> Object

It just bothers me that all the data types that make use of 'Object' have to either be declared in the "Dao.Evaluator" module or themselves be polymorphic. It feels like such a hack. I have decided to just make "Dao.Object" the god-module. I know it is bad design, but I think it is worse design to force polymorphism on a bunch of other data types that do not really need to be polymorphic just to avoid a module from getting to big.