Three bug fixes:
(1) Fixed the "delete()" function which I had stupidly defined to
automatically dereference parameter arguments, so it was not deleting
the references passed as arguments, rather it was trying to delete the
values stored at those references. This has been fixed now.

(2) The instance of the 'Dao.Interpreter.updateForLoop' function for the
'Dao.Interpreter.Object' data type was making use of the
'Dao.Interpreter.updateForLoopWith' which was using 'Data.Monoid.mempty'
as the initial value in the fold operation. But the
'Dao.Interpreter.Object' instance of 'Data.Monoid.mempty' is
'Dao.Interpreter.ONull' which was being treated as a single value rather
than an updated list, which was causing a null value to appear in lists
updated by "for" loop evaluation. By modifying the API for
'Dao.Interpreter.updateForLoopWith' and 'Dao.Interpreter.updateForLoop'
to not use 'Data.Monoid.mempty', and instead receive the initial value
as an argument, it becomes easier to customize the instantiation of
'Dao.Interpreter.updateForLoop' for the 'Dao.Interpreter.Object' data
type such that it behaves correctly without requiring a newtype for the
sole purpose of to instantiating 'Data.Monoid.Monoid' in a way that
makes the instantiation of 'Dao.Interpreter.UpdateIterable' work
correctly.

(3) The "for" loop expression should be able to modify the contents of
the object over which it is iterating by modifying the iterator
variable. While the 'Dao.Interpreter.updateForLoop' function was
correctly updating the list being iterated after fix (2), but the target
object of the iteration was not being updated after the loop completed.
I added a new function similar to 'derefObject' called
'derefObjectGetQualRef' which lets me update the 'Dao.Iterator.QualRef'
value associated with the target object of the iteration if any.