MAJOR BUG FIX!!!
This was a NON-TRIVIAL modification to the reference model of the Dao
interpreter that set me back three weeks. The previous reference model
did not allow for assigning to reference indices, so programming a
statement like:

    a[1] = 0;

was impossible . In fact, this feature is still broken, but now the
reference model at least allows for it to work. The actual fix involved
completely re-defining the 'Dao.Interpreter.Reference' object and the
related abstract syntax tree. I also created a new
'Dao.Interpreter.RefReducible' class as a counterpart to
'Dao.Interpreter.Executable', the difference is that
'Dao.Interpreter.RefReducible' always evaluates to a
'Dao.Interpreter.Reference', rather than completely executing it. So the
'Dao.Interpreter.Reference' data type now serves as kind of a thunk
object which can delay execution by the 'Dao.Interprter.execute'
function.

This modification not only allows for reference expressions to be used
to update the values they point to, they can also be used as a kind of
partial function-application. Partial application could become a feature
of the Dao programming language sometime in the future. It may also be
possible in the future to make tuples a built-in type, and write a tuple
of references on the left-hand side of an assignment expression.