Made a few more modifications to the reference model of the Dao programming language runtime. Now, all function calls are considered to be update operations, rather than lookup operations as they were before. I have created the 'Dao.Interpreter.refSuffixHasFuncCall' and 'Dao.Interpreter.refHasFuncCall' predicate functions to determine whether a 'Dao.Interpreter.Reference' value is a function call, and if it is, perform an update rather than a lookup.
This also made it necessary to modify the reference update behavior for
the 'Dao.Interpreter.ConstantStore'. Because all built-in functions are
stored in the 'Dao.Interpreter.ConstantStore' and since all function
calls perform updates, an update must be allowed to occur in the
'Dao.Interpreter.ConstantStore' without throwing an exception. What
really happens is, if an update is performed an update occurs on a copy
of the value in the store, then the store silently discards the updated
copy.