I have added two more built-in functions "toHash" and "fromHash", which allows a hashed Struct (in the Dao programming language, any data type defined with a label prefixed with a # character) to be converted to a built-in data type, and allows built-in data types to be converted to hashed Struct objects.
I have also added the 'Dao.Interpreter.objInterfaceName' field to the
'Dao.Interpreter.Interface' data type so the name of built-in data type
can be inferred as the type of a built-in object in the Dao runtime.
This required a modification to the 'Dao.Interpreter.TypeSym' structure
to create type symbols out of 'Dao.String.Name's.

Also, I have modified the behavior of 'Dao.Interpreter.coreType' and
'Dao.Interpreter.typeOfObj', and created a function
'Dao.Interpreter.objTypeFromCoreType' which does what
'Dao.Interpreter.coreType' used to do.