Changed the name of the 'Dao.Interpreter.GlobAction' data type to 'Dao.Interpreter.RulePattern'. I have also modified it such that this data type now store the pattern expressions in their literal form until the 'Dao.Interpreter.GlobAction' is ready to be stored into a 'Dao.Interpreter.RuleSet' or 'Dao.Glob.PatternTree'.
The 'Dao.Interpreter.RuleSet' has been modified to now be able to define
a customized string tokenizer.

The tokenizer data type, 'Dao.Interpreter.ExecTokenizer' now takes a
single string and returns a list of 'Dao.Interpreter.Object's which can
be strings, or any object that could be matched to the
'Dao.Interpreter.RulePattern's defined to use them.