Bug fix: rules with multiple patterns were being constructed such that each pattern was concatenated, rather than each pattern being converted to a glob and all globs being unioned together. This has been fixed so now the rule expressions construct rules in compliance with the official Dao programming language official specification that exists in my imagination.
To fix this bug, I instantiated 'Dao.Interpreter.PatternRule' into the
'Dao.Interpreter.Executable' class such that
'Dao.Interpreter.execute'ing a 'Dao.Interpreter.PatternRule' will
construct a pattern tree. This

Also, the other day I had eliminated the 'Dao.Interpreter.RefReducible'
class and hand-coded the special cases where that class was necessary.
But now, rather than hand coding, I have a new solution to the problem
of deciding when to evaluate a 'Dao.Interpreter.AST.AssignExpr' to a
'Dao.Interpreter.Reference' literal, and when to dereference that
'Dao.Interpreter.Reference' literal and return the value at the address
to which the 'Dao.Interpreter.Reference' is pointing.

The solution is to create a newtype wrapper for
'Dao.Interpreter.AST.AssignExpr' called
'Dao.Interpreter.DerefAssignExpr' which, when
'Dao.Interpreter.execute'd will automatically do the right thing,
dereferencing 'Dao.Interpreter.Reference' values to exactly one layer of
indirection, or zero layers if execution of the
'Dao.Interpreter.AST.AssignExpr' returned a result of an assignment
rather than a literal expression.