Minor bug fixes: the 'Dao.Interpreter.makeActionsForQuery' function was taking the list of strings that matched a typed glob wildcard and dumping the entire list into the parameter stack of the function call to the type check function, rather than enclosing the list of strings into a single list object and passing that list as the only parameter to the function call, causing type checking functions to backtrack due to incorrect arity of the function call. This has been fixed.
Also, the instantiation of
@('Dao.Interpreter.XPure' 'Dao.Interpreter.Object')@
into the 'Prelude.Ord' class was not converting the types of compared
object values to similar data types, so for example
@('Dao.Interpreter.OInt' 0) would ALWAYS be less than
@('Dao.Interpreter.OLong' 'Prelude.minBound')@. This was causing some
odd results in other arithmetic functions which make use of this
instantiation, namely the 'Prelude.^^' operator. This has been fixed.

Finally, modified the 'Dao.Interpreter.ExecInfixOpError' error and the
'Dao.Interpreter.typeMismatchError' function to throw an
'Dao.Interpreter.ExecInfixOpError' error.