Bug fix: glob patterns that ended with a single-token variable followed by a multi-token variable were not matching patterns correctly. For example the pattern: 	"$A?$B" The "$A?" variable (with a question mark, which indicates a single-token match) followed by the "$B" variable (with no question mark, which indicates a multi-token variable) should match any string of tokens with one token being assigned to A and zero or more tokens being assigned to B.
The bug was that if no tokens were assigned to B, i.e. B was to be an
empty list of tokens, the pattern would fail to match. This has been
fixed. Now B will be assigned an empty list of tokens if the above
example pattern were to be matched to a string of one tokens.

While I was at it, I also updated some of the comments, and renamed a
few of the type variables to be less confusing. Because as I was
trying to solve this bug I became confused by my variable names, I
thought I would name them more appropriately.