Bug fix: the "break" and "continue" expressions in the Dao programming language had not been working correctly, in fact I had not updated their behavior in years in spite of several changes made over the years that have rendered their original implementation obsolete. The current implementation now works correctly. I have added a test program to test the behavior of "break" and "continue" statements, the "tests/break-continue.dao" program.
Just as a note: the new implementation of the "break" and "continue"
statements borrows directly from the Python language in that
exceptions are thrown to break a loop. Like in the Python language,
break-exceptions do not travel beyond the boundaries of the current
function calling context. Unlike in the Python language, break
exceptions cannot be caught in the Dao program itself, it is strictly
internal to the Dao language interpreter.