I had to rewrite several of the instances and supporting functions which were defined to work in a polymorphic Monad "m" in the "Dao.Logic" and "Dao.Rule" monad to eliminate Functor and Applicative constraints "m". Having additional constraints like Functor or Applicative was preventing me from instantiating Rule into the MonadTrans class, which would be a major inconvenience to people making use of "Dao.Rule".
Do you know what I HATE about Haskell? That even if the context of a
polymorphic type includes a class (e.g. Monad) that is explicitly a
superclass of another subclass class (e.g. Functor), you are STILL not
fucking allowed to use the subclass functions for the polymorphic type
unless you include the subclass in the context as well. This means
when you have written all your nice, somewhat elegant code to make use
of Functor and Applicative, you have to declare the the polymorphic
type to be a member of Applicative and Functor even though it is
already a type of Monad just to make use of the Functor and
Applicative functions. But if you do this, and you try to instantiate
MonadTrans (which declares that the only restraint on the type is that
it be a member of Monad and not Functor or Applicative), it means you
cannot make use of any of your other code which makes use of Functor
or Applicative. So you have to fucking rewrite ALL of your functions
to makes sure they ONLY use Monad functions and nothing else, which
means you have to have a whole lot more do notation and a whole lot
more return statements just to accomplish the same thing that you had
previously accomplished with more elegant use of Functor and
Applicative, even though Functor and Applicative are subclasses of
Monad. Who's fucking bright idea was this?