I got to thinking about the naming of the 'Dao.Object.toForeign' and 'Dao.Object.fromForeign' functions and decided the names should be reversed. The 'Dao.Object.Foreign' data type has been renamed to the 'Dao.Object.Passport' data type because it is a wrapper around a foreign type, not the foreign type itself, and this was confusing me. The 'toForeign' function was originally the function used to wrap a foreign type to a 'Foreign' data type, and 'fromForeign' retrieved the original wrapped foreign data from the 'Foreign' data type -- this seemed to make sense because the container type we were converting from/to was called the 'Foreign' type.
What was so confusing was that the "from object to foreign type"
operation was performed with the 'fromForeign' function, and the "from
foreign type to object" operation was performed by the "toForeign"
function. Now that the 'Foreign' data type is called the 'Passport'
type, I was able to swap the names 'fromForeign' and 'toForeign' --
'fromForeign' is a function from a foreign type to a 'Passport', and
'toForeign' is a function from a 'Passport' to a foreign type.