Since newer versions of Haskell will now have the (&) operator in Prelude, which I think is a bad idea because it is defined simply as @flip ($)@ and is therefore rather redundant, none the less I must yield to popular decision. I have renamed the 'Dao.Lens.&' operator to 'Dao.Lens.~>'. I know this conflicts with some operator in the excellent pipes package, but I don't care. I use this operator far too often to make it an obscure three-character-long symbol that doesn't conflict with any popular packages.
I also renamed the 'Dao.Lens.on' function to 'Dao.Lens.with'. I knew
the name "on" conflicted with 'Dao.Function.on', and I thought to just
let the conflict exist. But since "Dao.Function" is in the "base"
package, I have decided to relent on this one too. Although the name
"with" is already used by 'Foreign.Marshal.Utils.with', I think people
using 'Foreign.Marshal.Utils.with' are working at a sufficiently low
level that they would not ordinarily import 'Dao.Lens's in the same
module where they would import "Foreign.Marshal.Utils". So there is
less conflict calling this lens function 'Dao.Lens.with' than there
would be calling it 'Dao.Lens.on'.