Renamed the constructors of the 'Dao.Predicate.Predicate' data type as the following:
1. OK        -> PTrue
2. Backtrack -> PFalse
3. PFail     -> PError

The "Dao.Predicate" module is the oldest code in the Dao package. It
has survived three or four near-total rewrites of the Dao package.
Originally it was intended to be used for parsing, but I found it
extremely useful to have a simple monadic function type was like
'Prelude.Maybe' but with the ability to also instantiate
'Control.Monad.Except.MonadError' (which at the time the
"Dao.Predicate" module was conceived, was called
'Control.Monad.Error').

Although I changed the name of the data type from 'Parser' to
'Predicate', the constructors 'OK', 'Backtrack', and 'PFail' remained
the same. The name 'Backtrack' is especially particular to parsing.
But it is no longer a parser, it is a predicate.

The name predicate, in the parlance of functional languages, means a
function that evaluates to a boolean True or False value. Likewise,
the 'Predicate' class is a monadic function type that can evaluate to
true or false. The only difference between a traditional predicate
function type and the Dao package's predicate monadic function type is
that 'Dao.Predicate.Predicate' has two additional features: it can
evaluate to a non-false error value, and the true value can be
accompanied with a return value. So to make the
'Dao.Predicate.Predicate' function a genuine predicate function type,
I have renamed the constructors to 'PTrue', 'PFalse,' and 'PError'.