Bug fix: I am a huge idiot with some kind of severe dyslexia or something. So, in the "Dao.Array" module, in defining the instance of 'Dao.TestNull.testNull' for the 'Dao.Array.Array' data type, I actually defined this function as 'Data.Maybe.isJust' rather than 'Data.Maybe.isNothing'. When the hell is 'Just' ever a null value 'Nothing' is non-null? What was I thinking?
Also in the "Dao.Interval" module, I had defined the
'Dao.Interval.intervalEnumSize' and 'Dao.Interval.intervalIntSize'
functions both to use 'Prelude.pred' instead of 'Prelude.succ', which
is only the hundredth time I have made this exact same mistake,
because I am dyslexic or something.

I have also modified the 'Dao.Text.CharSet.csetDecompose' function to
choose whether to invert a set before decomposing it based on the
'Dao.Interval.enumSize' of the set, rather than the maximum contiguous
interval as it was before. I am not sure exactly what I was thinking
when I wrote this function the first time around. Anyway, the idea is
to decompose the set into the least number of
'Dao.Interval.Interval's, and invert the set if the inversion produces
fewer 'Dao.Interval.Interval's, which is what just it does now.